Conflicts:
	nds-util/src/main/java/br/com/abril/nds/util/DateUtil.java

+++ nds-model/src/main/java/br/com/abril/nds/dto/filtro/FiltroEmissaoCE.java
@@ -17,9 +17,9 @@ public class FiltroEmissaoCE implements Serializable {
 	@Export(label="Até:")
 	private Date dtRecolhimentoAte;
 	@Export(label=" Intervalor Box:")
-	private Long idBoxDe;
+	private Integer idBoxDe;
 	@Export(label="Até:")
-	private Long idBoxAte;
+	private Integer idBoxAte;
 	@Export(label="Cota:")
 	private Integer numCotaDe;
 	@Export(label="Até:")
@@ -89,25 +89,25 @@ public class FiltroEmissaoCE implements Serializable {
 	/**
 	 * @return the idBoxDe
 	 */
-	public Long getIdBoxDe() {
+	public Integer getIdBoxDe() {
 		return idBoxDe;
 	}
 	/**
 	 * @param idBoxDe the idBoxDe to set
 	 */
-	public void setIdBoxDe(Long idBoxDe) {
+	public void setIdBoxDe(Integer idBoxDe) {
 		this.idBoxDe = idBoxDe;
 	}
 	/**
 	 * @return the idBoxAte
 	 */
-	public Long getIdBoxAte() {
+	public Integer getIdBoxAte() {
 		return idBoxAte;
 	}
 	/**
 	 * @param idBoxAte the idBoxAte to set
 	 */
-	public void setIdBoxAte(Long idBoxAte) {
+	public void setIdBoxAte(Integer idBoxAte) {
 		this.idBoxAte = idBoxAte;
 	}
 	/**
